+++ features/support/step_completions.rb
@@ -61,7 +61,7 @@ module StepCompletions
     click_button 'Continue'
   end
 
-  def complete_data_step
+  def complete_storing_step
     choose 'Just the team'
     @shared_duration = '1 year'
     @shared_usage = 'The data will be used to create better outcomes for more children'
@@ -71,7 +71,7 @@ module StepCompletions
     click_button 'Continue'
   end
 
-  def complete_time_equipment_step
+  def complete_where_when_step
     @held_on = '27th September at 2pm'
     fill_in 'The session is held on (optional)', with: @held_on
     @session_duration = '5 minutes'
@@ -91,7 +91,7 @@ module StepCompletions
     click_button 'Continue'
   end
 
-  def complete_incentive_step
+  def complete_incentives_step
     choose 'Yes'
 
     choose 'Cash'