+++ spec/models/research_session_spec.rb
@@ -173,8 +173,8 @@ RSpec.describe ResearchSession, type: :model do
         end
       end
 
-      describe 'validating the data step' do
-        let(:step) { :data }
+      describe 'validating the storing step' do
+        let(:step) { :storing }
 
         context 'no details are given' do
           let(:set_attrs) { { shared_with: nil, shared_duration: nil, shared_use: nil } }
@@ -205,8 +205,8 @@ RSpec.describe ResearchSession, type: :model do
         end
       end
 
-      describe 'validating the incentive step' do
-        let(:step) { :incentive }
+      describe 'validating the incentives step' do
+        let(:step) { :incentives }
 
         context 'no incentives are given, and no-one cares' do
           let(:set_attrs) { { incentive: nil } }