+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [FEATURE] - [Add location input to where&when](https://trello.com/c/wdLqH6qT/228-1-add-location-input-to-wherewhen)
 * [FEATURE] - [Create page titles for each step](https://trello.com/c/iZY1S1ZC/200-3-create-page-titles-for-each-step)
+* [FEATURE] - [Merge Topic / Purpose and update copy](https://trello.com/c/YIQw4isH/208-3-merge-topic-purpose-and-update-copy)
 
 # 0.3.0 / 2017-10-30